move parameters.scala to top/
