+++ ventus/src/pipeline/MSHR.scala
@@ -10,7 +10,7 @@
  * See the Mulan PSL v2 for more details. */
 package pipeline
 
-import parameters._
+import top.parameters._
 import chisel3._
 import chisel3.util._
