+++ ventus/src/pipeline/mul_utils.scala
@@ -12,7 +12,7 @@ package pipeline
 
 import chisel3._
 import chisel3.util._
-import pipeline.parameters._
+import top.parameters._
 object LookupTree {
   def apply[T <: Data](key: UInt, mapping: Iterable[(UInt, T)]): T =
     Mux1H(mapping.map(p => (p._1 === key, p._2)))
