+++ ventus/src/top/ExtMem_gen.scala
@@ -11,7 +11,7 @@
 package top
 import L1Cache.MyConfig
 import L2cache.{CacheParameters, InclusiveCacheMicroParameters, InclusiveCacheParameters_lite}
-import pipeline.parameters.num_warp
+import parameters.num_warp
 
 object GPGPU_gen extends App{
   (new chisel3.stage.ChiselStage).emitVerilog(new GPGPU_axi_top(),Array("--emission-options=disableMemRandomization,disableRegisterRandomization"))
